# backend/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Hypermedia19 API",
    "description" : "API for access MongoDB data for hypermedia project",
    "contact" : {
      "email" : "marco3.selva@mail.polimi.it"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/narcos9612/Hypermedia2019Festival/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "https://hypermedia19.herokuapp.com",
    "description" : "API use in project of hypermedia 19 in Politecnico di Milano. Gruop Selva, Tresoldi, Zani"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/order/{userId}/{eventId}" : {
      "get" : {
        "tags" : [ "authenticatedUsers" ],
        "summary" : "get Cart (order) of the user passed by Id",
        "description" : "You can get the  Cart (order) of the user passed by Id\n",
        "operationId" : "getCattBySserId",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "ID of user order",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "description" : "ID of event order",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "returnt the Single Order of the event and user passed by Id"
          },
          "404" : {
            "description" : "not found"
          },
          "500" : {
            "description" : "internal error"
          }
        }
      }
    },
    "/order/{userId}" : {
      "get" : {
        "tags" : [ "authenticatedUsers" ],
        "summary" : "get Cart (order) of the user passed by Id",
        "description" : "You can get the  Cart (order) of the user passed by Id\n",
        "operationId" : "getCartByUserId",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "ID of user order",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "the  of the user passed by Id",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrdersItem"
                }
              }
            }
          },
          "404" : {
            "description" : "eventId not found"
          },
          "500" : {
            "description" : "internal error"
          }
        }
      },
      "post" : {
        "tags" : [ "authenticatedUsers" ],
        "summary" : "post single order of the user passed by Id",
        "description" : "posting single order (new or update) and save in userId cart\n",
        "operationId" : "postOrderByUserId",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "ID of user order",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SingleOrderItem"
              },
              "example" : {
                "eventId" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
                "price" : 50,
                "quantity" : 2
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "succes of update or create single order",
            "content" : {
              "application/json" : {
                "example" : {
                  "message" : "order created"
                }
              }
            }
          },
          "404" : {
            "description" : "eventId not found"
          },
          "500" : {
            "description" : "internal error"
          }
        }
      },
      "delete" : {
        "tags" : [ "authenticatedUsers" ],
        "summary" : "delete  order of userId passed in params",
        "description" : "You can delete  artist with id passed in params\n",
        "operationId" : "deleteCartById",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "ID of userId that want to delete cart",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "delete event in the DB with id passed",
            "content" : {
              "application/json" : {
                "example" : {
                  "message" : "Order deleted",
                  "request" : {
                    "type" : "POST",
                    "url" : "https://hypermedia19.herokuapp.com/order/"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "unauthorized"
          },
          "500" : {
            "description" : "Cast error, not valid ObjectId"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "OrdersItem" : {
        "required" : [ "SingleOrder", "id", "totalPrice", "userId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "SingleOrder" : {
            "type" : "array",
            "items" : {
              "type" : "object"
            }
          },
          "totalPrice" : {
            "type" : "number",
            "example" : 50
          }
        }
      },
      "SingleOrderItem" : {
        "required" : [ "eventId", "id", "quantity", "subTotal", "userId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "quantity" : {
            "type" : "number",
            "example" : 3
          },
          "subTotal" : {
            "type" : "number",
            "example" : 50
          }
        }
      }
    },
...
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}